Move motor config classes to dedicated modules
